commit review refactoring code
