+++ apps/nextjs-website/react-components/components/Hero/Hero.helpers.tsx
@@ -1,0 +1,9 @@
+export const getMinHeight = (size: 'medium' | 'big' | 'small' | undefined) =>
+  size === 'big' ? '720px' : size === 'medium' ? '480px' : '220px';
+
+export const getOverlay = (useHoverlay: boolean, theme: string) =>
+  useHoverlay
+    ? theme === 'dark'
+      ? 'linear-gradient(0deg, rgba(0, 98, 195, 0.65), rgba(0, 98, 195, 0.65)), '
+      : 'linear-gradient(0deg, rgba(255, 255, 255, 0.8), rgba(255, 255, 255, 0.8)), '
+    : '';
