--- apps/nextjs-website/react-components/types/components.ts
@@ -1,11 +1,0 @@
-export type Theme = Readonly<'dark' | 'light'>;
-
-/** this Generic type is meant to be used
- * fot those components that will accept any JSX to
- * be rendered
- */
-export type Generic = JSX.Element;
-
-export interface CommonProps {
-  readonly theme: Theme;
-}
